+++ pkg/db/client.go
@@ -347,26 +347,26 @@ func (client *Client) SaveEvents(ctx context.Context, events []model.Event) erro
 	return nil
 }
 
-// SaveUserAgents implements the Store interface.
-func (client *Client) SaveUserAgents(ctx context.Context, userAgents []model.UserAgent) error {
+// SaveRequests implements the Store interface.
+func (client *Client) SaveRequests(ctx context.Context, requests []model.Request) error {
 	tx, err := client.Begin()
 
 	if err != nil {
 		return err
 	}
 
-	query, err := tx.PrepareContext(ctx, `INSERT INTO "user_agent" (time, user_agent) VALUES (?,?)`)
+	query, err := tx.PrepareContext(ctx, `INSERT INTO "request" (client_id, visitor_id, time, user_agent, path, event_name, bot) VALUES (?,?,?,?,?,?,?)`)
 
 	if err != nil {
 		return err
 	}
 
-	for _, ua := range userAgents {
-		_, err := query.Exec(ua.Time, ua.UserAgent)
+	for _, req := range requests {
+		_, err := query.Exec(req.ClientID, req.VisitorID, req.Time, req.UserAgent, req.Path, req.Event, req.Bot)
 
 		if err != nil {
 			if e := tx.Rollback(); e != nil {
-				client.logger.Error("error rolling back transaction to save user agents", "err", err)
+				client.logger.Error("error rolling back transaction to save requests", "err", err)
 			}
 
 			return err
@@ -378,43 +378,7 @@ func (client *Client) SaveUserAgents(ctx context.Context, userAgents []model.Use
 	}
 
 	if client.debug {
-		client.logger.Debug("saved user agents", "count", len(userAgents))
-	}
-
-	return nil
-}
-
-func (client *Client) SaveBots(ctx context.Context, bots []model.Bot) error {
-	tx, err := client.Begin()
-
-	if err != nil {
-		return err
-	}
-
-	query, err := tx.PrepareContext(ctx, `INSERT INTO "bot" (client_id, visitor_id, time, user_agent, path, event_name) VALUES (?,?,?,?,?,?)`)
-
-	if err != nil {
-		return err
-	}
-
-	for _, bot := range bots {
-		_, err := query.Exec(bot.ClientID, bot.VisitorID, bot.Time, bot.UserAgent, bot.Path, bot.Event)
-
-		if err != nil {
-			if e := tx.Rollback(); e != nil {
-				client.logger.Error("error rolling back transaction to save bots", "err", err)
-			}
-
-			return err
-		}
-	}
-
-	if err := tx.Commit(); err != nil {
-		return err
-	}
-
-	if client.debug {
-		client.logger.Debug("saved bots", "count", len(bots))
+		client.logger.Debug("saved requests", "count", len(requests))
 	}
 
 	return nil